Move DeveloperProject in CorePlugin
