+++ plugins/de.cognicrypt.core/src/de/cognicrypt/client/DeveloperProject.java
